### 2021-07-19 #3491 @fabian-s
@lutzhamel:

Your package currently does not follow many of the coding standards JOSS aims for (see below for the full output of `goodpractice::gp("popsom")`)  -- most importantly:

- please add automated tests with decent coverage for the core functionality
- please add proper dependency management ("Imports:", "Suggests:", ...) and namespace directives to the package

In addition, before we try to recruit reviewers, please:

- add the missing DOIs (see whedon bot output above)  to your paper
- use a suitable LICENSE for the package
- add a CONTRIBUTING file and/or community guidelines to the repository
(see also https://joss.readthedocs.io/en/latest/review_criteria.html)

```
> goodpractice::gp("popsom")
[...]
It is good practice to

  ✖ write unit tests for all functions, and all package code in
    general. 0% of code lines are covered by test cases.

    R/map-utils.R:67:NA
    R/map-utils.R:68:NA
    R/map-utils.R:70:NA
    R/map-utils.R:71:NA
    R/map-utils.R:73:NA
    ... and 1048 more lines

  ✖ write short and simple functions. These functions have high
    cyclomatic complexity:compute.centroids (98).
  ✖ omit "Date" in DESCRIPTION. It is not required and it gets
    invalid quite often. A build date will be added to the package when
    you perform `R CMD build` on it.
  ✖ add a "BugReports" field to DESCRIPTION, and point it to a
    bug tracker. Many online code hosting services provide bug trackers
    for free, https://github.com, https://gitlab.com, etc.

    R/map-utils.R:18:1
    R/map-utils.R:26:1
    R/map-utils.R:27:1
    R/map-utils.R:34:1
    R/map-utils.R:241:1
    ... and 27 more lines

  ✖ omit trailing semicolons from code lines. They are not
    needed and most R coding standards forbid them

    R/map-utils.R:256:32

    vapply() instead.

    R/map-utils.R:73:12
    R/map-utils.R:559:22
    R/map-utils.R:595:25
    R/map-utils.R:1640:12
    R/map-utils.R:1759:22

  ✖ avoid the library() and require() functions, they change
    the global search path. If you need to use other packages, import
    them. If you need to load them explicitly, then consider
    loadNamespace() instead, or as a last resort, declare them as
    'Depends' dependencies.

    R/map-utils.R:39:1
    R/map-utils.R:40:1
    R/map-utils.R:41:1
    R/map-utils.R:42:1

  ✖ avoid 1:length(...), 1:nrow(...), 1:ncol(...), 1:NROW(...)
    and 1:NCOL(...) expressions. They are error prone and result 1:0 if
    the expression on the right hand side is zero. Use seq_len() or
    seq_along() instead.

    R/map-utils.R:406:15
    R/map-utils.R:475:13
    R/map-utils.R:492:22
    R/map-utils.R:495:15
    R/map-utils.R:549:22
    ... and 10 more lines

  ✖ not import packages as a whole, as this can cause name
    clashes between the imported packages. Instead, import only the
    specific functions you need.
```

Looking forward to working with you on this submission,
best,
Fabian
