### 2021-07-19 #3491 @lutzhamel
@fabian-s Interesting automated feedback for the package.  Quick question: The R release process requires you to embed vignettes into the documentation.  These vignettes are treated by the R release process as automated test cases and the release process will terminate if these cases do not succeed. Neither will the release process continue if the documentation does not include these vignettes.  Now, do I have to replicate this functionality with an additional test suite?
