### 2021-07-19 #3491 @fabian-s
@lutzhamel

I don't agree with the implied premise that vignettes, which are part of your documentation, are equivalent to regression or unit tests or can replace them.

I do concede that, by having vignettes that are being rebuilt whenever the package is bundled, you will frequently discover if something breaks really badly, but not in a systematic or structured way (e.g. you'll know about the first bug that breaks your vignette, but not subsequent ones. or, if a function still runs but returns a wrong result, the vignette will render just fine with nonsensical content).

Turning examples in your vignettes or help pages into unit tests can be a good starting point if they run quickly, cover the API of your package and each focus on separate functionalities -- for additional considerations on what to test and how, see for example section 7 in https://cran.r-project.org/web/packages/tinytest/vignettes/using_tinytest.pdf or [Cotton: Testing R Code](https://www.routledge.com/Testing-R-Code/Cotton/p/book/9780367782375) for more in-depth info.
