### 2021-07-19 #3491 @lutzhamel
@fabian-s Got it.  Thanks for the links, very helpful.  Another quick question, is there a way for me to run the gp() function manually in order to check that my code conforms to the joss standards?

Thanks,
Lutz
