### 2021-07-19 #3491 @fabian-s
Sure, glad it was helpful.

re. `gp()`:
not sure I really understand what you're asking -- to use it you install the `goodpractice` package from CRAN and then call its `gp()`-function with the folder path to the source R package you want to check.

Note that `gp()` does not define/implement JOSS standards for submitted R code, though, I didn't mean to create that impression -- it's just a quick and easy way to find some low-hanging fruit for possible improvements.

What JOSS is looking for in a paper is described at https://joss.readthedocs.io/en/latest/submitting.html and https://joss.readthedocs.io/en/latest/review_criteria.html.
