### 2021-07-19 #3491 @lutzhamel
Thanks for getting back to me so quickly.  I didn't realize that gp() was part of a CRAN package.  I will follow up.

Best,
Lutz
