+++ joss/paper.bib
@@ -19,7 +19,8 @@
   volume={119},
   pages={9--17},
   year={2018},
-  publisher={Elsevier}
+  publisher={Elsevier},
+  doi={10.1016/j.cageo.2018.06.006}
 }
 
 @article{mathys2019single,
@@ -52,7 +53,8 @@
   number={1},
   pages={293--300},
   year={1996},
-  publisher={Blackwell Science Ltd Oxford, UK}
+  publisher={Blackwell Science Ltd Oxford, UK},
+  doi={10.1093/mnras/279.1.293},
 }
 
  @Article{kohonen2018jss,
@@ -89,7 +91,8 @@ Maps (SOMs) with Starburst Visualizations},
   booktitle={Proceedings of SAI Intelligent Systems Conference},
   pages={805--821},
   year={2018},
-  organization={Springer}
+  organization={Springer},
+  doi={10.1007/978-3-030-01057-7_60}
 }
 
 @inproceedings{hamel2011improved,
@@ -116,7 +119,8 @@ Maps (SOMs) with Starburst Visualizations},
   booktitle={Advances in Self-Organizing Maps and Learning Vector Quantization},
   pages={49--59},
   year={2016},
-  publisher={Springer}
+  publisher={Springer},
+  doi={10.1007/978-3-319-28518-4_4}
 }
 
 @article{tatoian2018self,
