+++ DESCRIPTION
@@ -1,6 +1,5 @@
 Package: popsom
-Version: 5.2
-Date: 2021-7-9
+Version: 5.3
 Title: A Very Efficient Implementation of Kohonen's Self-Organizing Maps (SOMs) with Starburst Visualizations
 Authors@R: c(person("Lutz", "Hamel", role = c("aut", "cre"), email = "lutzhamel@uri.edu"),
              person("Benjamin", "Ott", role = "aut"),
@@ -14,5 +13,6 @@ Description: Kohonen's self-organizing maps with a number of distinguishing feat
     (2) Automatic centroid detection and visualization using starbursts.
     (3) Two models of the data: (a) a self-organizing map model, (b) a centroid based clustering model.
     (4) A number of easily accessible quality metrics for the self-organizing map and the centroid based cluster model.
-License: GPL
+License: GPL-3
 URL: https://github.com/lutzhamel/popsom
+BugReports: https://github.com/lutzhamel/popsom/issues
