### 2021-07-21 62db2a4
making package joss compliant
