+++ R/map-utils.R
@@ -15,31 +15,32 @@
 # position ------ return the position of points on the map
 # significance -- graphically reports the significance of each feature with
 #                 respect to the self-organizing map model
-# marginal ------ displays a density plot of a training dataframe dimension overlayed
-#                 with the neuron density for that same dimension or index.
+# marginal ------ displays a density plot of a training dataframe dimension
+#                 overlayed with the neuron density for that same dimension or
+#                 index.
 #
 ### License
 # This program is free software; you can redistribute it and/or modify it under
 # the terms of the GNU General Public License as published by the Free Software
 # Foundation.
 #
-# This program is distributed in the hope that it will be useful, but WITHOUT ANY
-# WARRANTY; without even the implied warranty of MERCHANTABILITY or FITNESS FOR A
-# PARTICULAR PURPOSE. See the GNU General Public License for more details.
+# This program is distributed in the hope that it will be useful, but WITHOUT
+# ANY WARRANTY; without even the implied warranty of MERCHANTABILITY or FITNESS
+# FOR A PARTICULAR PURPOSE. See the GNU General Public License for more details.
 #
 #  A copy of the GNU General Public License is available at
 #  http://www.r-project.org/Licenses/
 #
 ### Helpful docs
-# https://medium.com/@ODSC/unsupervised-learning-evaluating-clusters-bd47eed175ce
-# https://en.wikipedia.org/wiki/Total_sum_of_squares
+# medium.com/@ODSC/unsupervised-learning-evaluating-clusters-bd47eed175ce
+# en.wikipedia.org/wiki/Total_sum_of_squares
 ###
 
-# load libraries
-require(fields)
-require(graphics)
-require(ggplot2)
-require(hash)
+# load libraries -- avoid changing the search path
+loadNamespace("fields")
+loadNamespace("graphics")
+loadNamespace("ggplot2")
+loadNamespace("hash")
 
 ### constructor ###
 
@@ -238,7 +239,8 @@ starburst <- function(map)
 # parameters:
 # - map is an object if type 'map'
 # - graphics is a switch that controls whether a plot is generated or not
-# - feature.labels is a switch to allow the plotting of feature names vs feature indices
+# - feature.labels is a switch to allow the plotting of feature names vs
+#   feature indices
 # return value:
 # - a vector containing the significance for each feature
 significance <- function(map,graphics=TRUE,feature.labels=TRUE)
@@ -253,7 +255,7 @@ significance <- function(map,graphics=TRUE,feature.labels=TRUE)
 	var.v <- array(data=1,dim=nfeatures)
 	for (i in 1:nfeatures)
     {
-		var.v[i] <- var(data.df[[i]]);
+		var.v[i] <- var(data.df[[i]])
     }
 
 	# we use the variance of a feature as likelihood of
@@ -295,7 +297,8 @@ significance <- function(map,graphics=TRUE,feature.labels=TRUE)
 	}
 }
 
-# marginal - plot that shows the marginal probability distribution of the neurons and data
+# marginal - plot that shows the marginal probability distribution of the
+#            neurons and data
 # parameters:
 # - map is an object of type 'map'
 # - marginal is the name of a training data frame dimension or index
@@ -315,7 +318,9 @@ marginal <- function(map,marginal)
     neurons$legend <- 'neurons'
 
     hist <- rbind(train,neurons)
[…]
