+++ CONTRIBUTING.md
@@ -1,3 +1,3 @@
 # Contributing
 
-We welcome suggestions and contributions. You could  either directly contact us via [email](mailto:lutz.hamel@gmail.com) or you can create an issue in our [GitHub repository](https://github.com/lutzhamel/popsom). Please do not submit a pull request without prior consultation with us.
+We welcome suggestions and contributions. You could  either directly contact us via [email](mailto:lutzhamel@uri.edu) or you can create an issue in our [GitHub repository](https://github.com/lutzhamel/popsom). Please do not submit a pull request without prior consultation with us.
