+++ package/tests/testthat/test-map.R
@@ -1,0 +1,9 @@
+test_that("map constructors work", {
+  data(iris)
+  df<-subset(iris,select=-Species)
+  labels<-subset(iris,select=Species)
+  m<-map(df,labels,seed=42)
+
+  expect_equal(m$convergence < 0.8,TRUE)
+  expect_equal(length(m$unique.centroids)<5,TRUE)
+})
