+++ package/tests/testthat.R
@@ -1,0 +1,4 @@
+library(testthat)
+library(popsom)
+
+test_check("popsom")
