+++ package/DESCRIPTION
@@ -8,6 +8,8 @@ Authors@R: c(person("Lutz", "Hamel", role = c("aut", "cre"), email = "lutzhamel@
              person("Michael", "Eiger", role = "aut"),
              person("Vishakh", "Gopu", role = "aut"))
 Imports: fields, graphics, ggplot2, hash, stats, grDevices
+Suggests: 
+    testthat (>= 3.0.0)
 Description: Kohonen's self-organizing maps with a number of distinguishing features:
     (1) A very efficient, single threaded, stochastic training algorithm based on ideas from tensor algebra.  Up to 60x faster than traditional single-threaded training algorithms. No special accelerator hardware required.
     (2) Automatic centroid detection and visualization using starbursts.
@@ -16,3 +18,4 @@ Description: Kohonen's self-organizing maps with a number of distinguishing feat
 License: GPL-3
 URL: https://github.com/lutzhamel/popsom
 BugReports: https://github.com/lutzhamel/popsom/issues
+Config/testthat/edition: 3
