### 2021-07-21 fc38dfb
changed library calls to loadNamespace calls
