### 2021-07-21 #3491 @lutzhamel
@fabian-s

I believe I have addressed all the issues raised and adjusted the repository appropriately.

* please add proper dependency management ("Imports:", "Suggests:", ...) and namespace directives to the package
>>> done.  The DESCRIPTION file has an "Imports:" section for the R package manager.

* add the missing DOIs (see whedon bot output above) to your paper
>>> done.

* use a suitable LICENSE for the package
>>> done -- GPL-3

* add a CONTRIBUTING file and/or community guidelines to the repository
>>> done

* please add automated tests with decent coverage for the core functionality
✖ write unit tests for all functions, and all package code in
    general. 0% of code lines are covered by test cases.

    R/map-utils.R:67:NA
    R/map-utils.R:68:NA
    R/map-utils.R:70:NA
    R/map-utils.R:71:NA
    R/map-utils.R:73:NA
    ... and 1048 more lines
>>> Done. My unit tests cover about 75% of the code.

✖ write short and simple functions. These functions have high
    cyclomatic complexity:compute.centroids (98).
>>> I did not follow this advice.  The function mentioned has to test A LOT of boundary conditions and therefore
the cyclomatic complexity is high.  Breaking the function into smaller parts
would render the algorithm behind the function virtually incomprehensible.

✖ omit "Date" in DESCRIPTION. It is not required and it gets
    invalid quite often. A build date will be added to the package when
    you perform `R CMD build` on it.
>>> done

✖ add a "BugReports" field to DESCRIPTION, and point it to a
    bug tracker. Many online code hosting services provide bug trackers
    for free, https://github.com, https://gitlab.com, etc.
>>> done

    R/map-utils.R:18:1
    R/map-utils.R:26:1
    R/map-utils.R:27:1
    R/map-utils.R:34:1
    R/map-utils.R:241:1
    ... and 27 more lines
>>> done

✖ omit trailing semicolons from code lines. They are not
    needed and most R coding standards forbid them

    R/map-utils.R:256:32
>>> Done

    vapply() instead.

    R/map-utils.R:73:12
    R/map-utils.R:559:22
    R/map-utils.R:595:25
    R/map-utils.R:1640:12
    R/map-utils.R:1759:22
>>> I did not follow this advice.  R is not type safe,  see
https://stackoverflow.com/questions/14298899/type-safety-of-the-r-language#:~:text=R%20has%20essentially%20no%20type,end%20up%20with%20meaningless%20objects for examples. However, R is dynamically typed.  The advantage of dynamically typed languages is the readability
of the code. `sapply` is the most readable and intuitive of the `apply` family of functions in R.
Why should I sacrifice this readability for the pretense of type safety?

✖ avoid the library() and require() functions, they change
    the global search path. If you need to use other packages, import
    them. If you need to load them explicitly, then consider
    loadNamespace() instead, or as a last resort, declare them as
    'Depends' dependencies.

    R/map-utils.R:39:1
    R/map-utils.R:40:1
    R/map-utils.R:41:1
    R/map-utils.R:42:1
>>> Done

✖ avoid 1:length(...), 1:nrow(...), 1:ncol(...), 1:NROW(...)
    and 1:NCOL(...) expressions. They are error prone and result 1:0 if
    the expression on the right hand side is zero. Use seq_len() or
    seq_along() instead.

    R/map-utils.R:406:15
    R/map-utils.R:475:13
    R/map-utils.R:492:22
    R/map-utils.R:495:15
    R/map-utils.R:549:22
    ... and 10 more lines
>>> I did not follow this advice. '1:length()' is an idiomatic expression.  Any programmer reading this
code immediately understands what is going on, e.g 'for (i in 1:length(x))...'
This is not true for seq_len.  The name is unfortunate and it is not
clear what 'for (i in seq_len(length(x)))...' really does unless you start
digging into the documentation.  The safety improvements are minimal and
do not outweigh the readability concerns of the code.

✖ not import packages as a whole, as this can cause name
    clashes between the imported packages. Instead, import only the
    specific functions you need.
>>> I did not follow this advice. There are no name clashes between the imported packages.  Should there
be in the future this will be addressed.
