### 2021-07-21 a9e92d5
adjusted the build instructions
