### 2021-07-22 #3524 @whedon
[REVIEW]: popsom: A Very Efficient Implementation of Self-Organizing Maps with Starburst Visualizations for R
**Submitting author:** @lutzhamel (<a href="http://orcid.org/0000-0002-2302-3358">Lutz Hamel</a>)
**Repository:** <a href="https://github.com/lutzhamel/popsom" target ="_blank">https://github.com/lutzhamel/popsom</a>
**Version:** v5.2
