### 2021-08-09 lutzhamel/popsom#7 @HerrMo
S3 generics get masked
Functions `popsom::summary`, `popsom::predict` and `popsom::fitted` mask S3 generics.

This is very inconvenient as it breaks method dispatch for objects of other classes when `popsom` is loaded.
Appropriate S3 methods should therefore be implemented.

With this in mind, I also suggest to move the `starburst` functionality into a S3 plot method and to add a S3 print method.
