### 2021-08-09 lutzhamel/popsom#9 @HerrMo
Performance claims cannot be confirmed (using simple experiments)
I have difficulties confirming the performance claims (see the experiments below). However, it may well be that the experiments performed are oversimplified and/or poorly specified (they are mainly based on default settings).
It would be very helpful if you could provide instructions and examples that can be used to test the performance. I also suggest to include an example illustrating the performance improvements in the software paper.

``` r
library(popsom)
library(som)
library(kohonen)
library(MASS)
library(microbenchmark)
library(ggplot2)

# method wrappers

pop_wrp <- function(dat, ...) popsom::map(as.data.frame(dat), ...)
som_wrp <- function(dat, ...) som::som(dat, ...)
koh_wrp <- function(dat, ...) kohonen::som(as.matrix(dat), ...)

# data sets

# iris
data(iris)
df_iris <- subset(iris, select = -Species)

# wines from package kohonen
data(wines)
df_wines <- scale(wines)

# synthetic data with three clusters
p <- 10
n <- 500
siglarg <- diag(rep(1, p * p), p, p)
means <- c(0, -50, 50)

clusts <- lapply(means, function(mu) mvrnorm(n = n, mu = rep(mu, p), Sigma = siglarg))
df_sim <- do.call(rbind, clusts)

datsets <- list(df_iris, df_wines, df_sim)
bmr <- lapply(datsets,
              function(dat) microbenchmark(pop_wrp(dat, train = 1000),
                                           som_wrp(dat, xdim = 10, ydim = 5), # no default values for xdim and ylim. set to popsom defaults
                                           koh_wrp(dat)))

ggplot2::autoplot(bmr[[1]]) + ggplot2::ggtitle("Wine data")
#> Coordinate system already present. Adding new coordinate system, which will replace the existing one.
```

![](https://i.imgur.com/EX007p4.png)

``` r
ggplot2::autoplot(bmr[[2]]) + ggplot2::ggtitle("Iris data")
#> Coordinate system already present. Adding new coordinate system, which will replace the existing one.
```

![](https://i.imgur.com/GOFt9d4.png)

``` r
ggplot2::autoplot(bmr[[3]]) + ggplot2::ggtitle("Simulated data")
#> Coordinate system already present. Adding new coordinate system, which will replace the existing one.
```

![](https://i.imgur.com/1lb0MCD.png)

``` r
sessionInfo()
#> R version 4.1.0 (2021-05-18)
#> Platform: x86_64-pc-linux-gnu (64-bit)
#> Running under: Linux Mint 19.2
#>
#> Matrix products: default
#> BLAS:   /usr/lib/x86_64-linux-gnu/blas/libblas.so.3.7.1
#> LAPACK: /usr/lib/x86_64-linux-gnu/lapack/liblapack.so.3.7.1
#>
#> locale:
#>  [1] LC_CTYPE=en_US.UTF-8       LC_NUMERIC=C
#>  [3] LC_TIME=en_US.UTF-8        LC_COLLATE=en_US.UTF-8
#>  [5] LC_MONETARY=de_DE.UTF-8    LC_MESSAGES=en_US.UTF-8
#>  [7] LC_PAPER=de_DE.UTF-8       LC_NAME=C
#>  [9] LC_ADDRESS=C               LC_TELEPHONE=C
#> [11] LC_MEASUREMENT=de_DE.UTF-8 LC_IDENTIFICATION=C
#>
#> attached base packages:
#> [1] stats     graphics  grDevices utils     datasets  methods   base
#>
#> other attached packages:
#> [1] ggplot2_3.3.5        microbenchmark_1.4-7 MASS_7.3-54
#> [4] kohonen_3.0.10       som_0.3-5.1          popsom_5.2
#>
#> loaded via a namespace (and not attached):
#>  [1] Rcpp_1.0.7        compiler_4.1.0    pillar_1.6.1      highr_0.9
#>  [5] viridis_0.6.1     tools_4.1.0       dotCall64_1.0-1   digest_0.6.27
#>  [9] viridisLite_0.4.0 evaluate_0.14     lifecycle_1.0.0   tibble_3.1.2
#> [13] gtable_0.3.0      pkgconfig_2.0.3   rlang_0.4.11      reprex_2.0.1
#> [17] cli_3.0.0         rstudioapi_0.13   yaml_2.2.1        spam_2.7-0
#> [21] xfun_0.24         gridExtra_2.3     withr_2.4.2       stringr_1.4.0
#> [25] dplyr_1.0.6       knitr_1.33        maps_3.3.0        fields_12.5
#> [29] generics_0.1.0    fs_1.5.0          vctrs_0.3.8       tidyselect_1.1.1
#> [33] grid_4.1.0        glue_1.4.2        R6_2.5.0          hash_2.2.6.1
#> [37] fansi_0.5.0       rmarkdown_2.8     farver_2.1.0      purrr_0.3.4
#> [41] magrittr_2.0.1    scales_1.1.1      htmltools_0.5.1.1 ellipsis_0.3.2
#> [45] colorspace_2.0-2  utf8_1.2.1        stringi_1.6.2     munsell_0.5.0
#> [49] crayon_1.4.1
```
