### 2021-08-09 lutzhamel/popsom#10 @HerrMo
Uninformative error message if learning rate is not adequately specified
The documentation of `map` states that `alpha`, "the learning rate, should be a positive non-zero real number."
However, a non-zero positive learning rate can result in an error with a rather uninformative error message (see MRE below).
I suggest to provide a more meaningful error message, for example -- if this is the case -- that the learning rate is chosen to large.

``` r
library(popsom)
data(iris)
df <- subset(iris, select = -Species)
labels <- subset(iris ,select = Species)

# triggers error
m <- popsom::map(df, labels, xdim = 15, ydim = 10, train = 10000, alpha = 2.3, seed = 42)
#> Error in ks.test(map.df[[i]], data.df[[i]]): not enough 'x' data

# does not trigger error
m <- popsom::map(df, labels, xdim = 15, ydim = 10, train = 10000, alpha = 2.2, seed = 42)
```

``` r
sessionInfo()
#> R version 4.1.0 (2021-05-18)
#> Platform: x86_64-pc-linux-gnu (64-bit)
#> Running under: Linux Mint 19.2
#>
#> Matrix products: default
#> BLAS:   /usr/lib/x86_64-linux-gnu/blas/libblas.so.3.7.1
#> LAPACK: /usr/lib/x86_64-linux-gnu/lapack/liblapack.so.3.7.1
#>
#> locale:
#>  [1] LC_CTYPE=en_US.UTF-8       LC_NUMERIC=C
#>  [3] LC_TIME=en_US.UTF-8        LC_COLLATE=en_US.UTF-8
#>  [5] LC_MONETARY=de_DE.UTF-8    LC_MESSAGES=en_US.UTF-8
#>  [7] LC_PAPER=de_DE.UTF-8       LC_NAME=C
#>  [9] LC_ADDRESS=C               LC_TELEPHONE=C
#> [11] LC_MEASUREMENT=de_DE.UTF-8 LC_IDENTIFICATION=C
#>
#> attached base packages:
#> [1] stats     graphics  grDevices utils     datasets  methods   base
#>
#> other attached packages:
#> [1] popsom_5.2
#>
#> loaded via a namespace (and not attached):
#>  [1] compiler_4.1.0       pillar_1.6.1         highr_0.9
#>  [4] viridis_0.6.1        tools_4.1.0          dotCall64_1.0-1
#>  [7] digest_0.6.27        viridisLite_0.4.0    evaluate_0.14
#> [10] lifecycle_1.0.0      tibble_3.1.2         gtable_0.3.0
#> [13] pkgconfig_2.0.3      rlang_0.4.11         reprex_2.0.1
#> [16] cli_3.0.0            rstudioapi_0.13      microbenchmark_1.4-7
#> [19] yaml_2.2.1           spam_2.7-0           xfun_0.24
#> [22] gridExtra_2.3        withr_2.4.2          stringr_1.4.0
#> [25] dplyr_1.0.6          knitr_1.33           maps_3.3.0
#> [28] fields_12.5          generics_0.1.0       fs_1.5.0
#> [31] vctrs_0.3.8          tidyselect_1.1.1     grid_4.1.0
#> [34] glue_1.4.2           R6_2.5.0             hash_2.2.6.1
#> [37] fansi_0.5.0          rmarkdown_2.8        purrr_0.3.4
#> [40] ggplot2_3.3.5        magrittr_2.0.1       scales_1.1.1
#> [43] htmltools_0.5.1.1    ellipsis_0.3.2       colorspace_2.0-2
#> [46] utf8_1.2.1           stringi_1.6.2        munsell_0.5.0
#> [49] crayon_1.4.1
```
