### 2021-08-09 lutzhamel/popsom#11 @HerrMo
Include Community guidelines and Installation instructions in the Readme
In my opinion, it would be good practice to include installation instructions in the Readme, for example:

### Installation

Install the last release from CRAN:

```
install.packages("popsom")
```

 Moreover, I think a note on how to contribute/report bugs is missing.
