### 2021-08-09 lutzhamel/popsom#12 @HerrMo
Include a brief explanation of the example in the software paper
The documentation and the included examples already make the use of `popsom` very easy.  To make the package even more accessible, I suggest including a brief explanation of the summary output and the starburst plot in section **Usage** of the software paper.
