### 2021-08-09 #3524 @HerrMo
@fabian-s @whedon I have finished my review and I suggest a revision as follows

Summary:
The package `popsom` provides an implementation of self-organizing maps. The software is well documented, easy to use and provides visualizations and quality metrics which make the assessment of a trained map very intuitive. The software paper is well written and gives an accurate and for the most part sufficient overview of the package and the problems it solves. However, several issues remain that require revision and/or clarification, thus blocking acceptance.

Major issues:
- https://github.com/lutzhamel/popsom/issues/7#issue-963989220
- https://github.com/lutzhamel/popsom/issues/8#issue-964056932
- https://github.com/lutzhamel/popsom/issues/9#issue-964151813

Minor issues:
- https://github.com/lutzhamel/popsom/issues/10#issue-964195440
- https://github.com/lutzhamel/popsom/issues/11#issue-964228333
- https://github.com/lutzhamel/popsom/issues/12#issue-964254134
