### 2021-08-09 lutzhamel/popsom#7 @lutzhamel
@HerrMo I had S3 interfaces implemented for R3.x and they worked beautifully.  But when R4.x was released the S3 interfaces stopped working.  There was no documentation on what changed from R3.x -> R4.x that would make S3 interfaces break.  I wanted POPSOM to work on both versions and therefore took the S3 interfaces out.

If you have any insights on what needs to happen for R4.x to implement S3 interfaces or what changed from R3.x to R4.x that would be greatly appreciated.  Otherwise I am afraid I only have two other options: (a) leave it as is or (b) rename the functions so they don't interfere with other S3 functions.  This package needs to be compatible with both R3.x and R4.x because a lot of Unix based systems still have R3.x as their default installation.
