### 2021-08-09 lutz-hamel/popsom#11 @lutzhamel
There is a file called CONTRIBUTING.md that details how to participate in this project.
