### 2021-08-11 lutz-hamel/popsom#8 @lutzhamel
Fixed.  In severely undertrained maps the whole map can look like a single cluster.  The original code did not accommodate for that.  This has been fixed now.
