### 2021-08-11 lutz-hamel/popsom#10 @lutzhamel
Fixed.  We now restrict alpha values to 0 < alpha <= 1 and report and appropriate error message if the value falls outside of that range.
