### 2021-08-11 lutz-hamel/popsom#7 @lutzhamel
I decided to sidestep the whole S3 issue by renaming the functions in the interface.  As far as I can tell there are now conflicts at this point.
