### 2021-08-11 lutz-hamel/popsom#9 @lutzhamel
I had to adjust the parameters of the various algorithms so that they all compute roughly the same thing.  Here is the rewritten script,
```
library(popsom)
library(som)
library(kohonen)
library(MASS)
library(microbenchmark)

# data sets

# iris
data(iris)
df_iris <- subset(iris, select = -Species)

# wines from package kohonen
data(wines)
df_wines <- as.data.frame(scale(wines))

# synthetic data with three clusters
p <- 10
n <- 500
siglarg <- diag(rep(1, p * p), p, p)
means <- c(0, -50, 50)

clusts <- lapply(means, function(mu) mvrnorm(n = n, mu = rep(mu, p), Sigma = siglarg))
df_sim <- as.data.frame(do.call(rbind, clusts))

# benchmark parameters
xdim <- 10
ydim <- 5
r <- sqrt(xdim**2+ydim**2)
iter <- 1000

# benchmark expressions
bm <- function(dat) {
  data.frame(summary(microbenchmark(
    popsom::map.build(dat,
                      xdim=xdim,
   	              ydim=ydim,
	              train = iter,
		      minimal=TRUE),

    som::som(dat,
             xdim = xdim,
	     ydim = ydim,
	     init="random",
             alpha=c(.3,.3),
             alphaType="linear",
             neigh="bubble",
             topol="rect",
             radius=c(r,r),
             rlen=c(1,iter)),

     kohonen::som(as.matrix(dat),
                  rlen=iter,
	          grid=somgrid(xdim=xdim,ydim=ydim,topo="rectangular",neighbourhood.fct="bubble")),

     times=3)))
}

# run the benchmarks

print("times are reported in milliseconds")

print("### Iris ###")
bmr <- bm(df_iris)
d <- data.frame(cbind(c("popsom","som","kohonen"),format(bmr[["mean"]],digits=2)))
names(d)<-c("package","mean time")
print(d)

print("### Wines ###")
bmr <- bm(df_wines)
d <- data.frame(cbind(c("popsom","som","kohonen"),bmr[["mean"]]))
names(d)<-c("package","mean time")
print(d)

print("### Sim ###")
bmr <- bm(df_sim)
d <- data.frame(cbind(c("popsom","som","kohonen"),bmr[["mean"]]))
names(d)<-c("package","mean time")
print(d)
```
The output is,
```
[1] "times are reported in milliseconds"
[1] "### Iris ###"
  package mean time
1  popsom       1.7
2     som      17.2
3 kohonen     128.6
[1] "### Wines ###"
  package        mean time
1  popsom         3.207418
2     som 21.8024186666667
3 kohonen       281.059903
[1] "### Sim ###"
  package        mean time
1  popsom 2.81184866666667
2     som 33.9963803333333
3 kohonen      1975.985098
```
which shows that `popsom` is about an order of magnitude faster than `som` and almost two orders of magnitude faster than `kohonen`.  The script is now part of the `popsom` release.

```
R version 4.1.0 (2021-05-18)
Platform: x86_64-pc-linux-gnu (64-bit)
Running under: Ubuntu 20.04.2 LTS

Matrix products: default
BLAS:   /usr/lib/x86_64-linux-gnu/blas/libblas.so.3.9.0
LAPACK: /usr/lib/x86_64-linux-gnu/lapack/liblapack.so.3.9.0

locale:
 [1] LC_CTYPE=C.UTF-8       LC_NUMERIC=C           LC_TIME=C.UTF-8
 [4] LC_COLLATE=C.UTF-8     LC_MONETARY=C.UTF-8    LC_MESSAGES=C.UTF-8
 [7] LC_PAPER=C.UTF-8       LC_NAME=C              LC_ADDRESS=C
[10] LC_TELEPHONE=C         LC_MEASUREMENT=C.UTF-8 LC_IDENTIFICATION=C

attached base packages:

other attached packages:
[1] ggplot2_3.3.5        microbenchmark_1.4-7 MASS_7.3-54
[4] kohonen_3.0.10       som_0.3-5.1          popsom_6.0

 [1] Rcpp_1.0.7        pillar_1.6.1      compiler_4.1.0    viridis_0.6.1
 [5] tools_4.1.0       dotCall64_1.0-1   lifecycle_1.0.0   tibble_3.1.2
 [9] gtable_0.3.0      viridisLite_0.4.0 lattice_0.20-44   pkgconfig_2.0.3
[13] rlang_0.4.11      Matrix_1.3-4      mvtnorm_1.1-2     spam_2.7-0
[17] gridExtra_2.3     withr_2.4.2       dplyr_1.0.7       generics_0.1.0
[21] vctrs_0.3.8       fields_12.5       maps_3.3.0        grid_4.1.0
[25] tidyselect_1.1.1  glue_1.4.2        R6_2.5.0          hash_2.2.6.1
[29] fansi_0.5.0       survival_3.2-11   multcomp_1.4-17   TH.data_1.0-10
[33] purrr_0.3.4       magrittr_2.0.1    codetools_0.2-18  splines_4.1.0
[37] scales_1.1.1      ellipsis_0.3.2    colorspace_2.0-2  sandwich_3.0-1
[41] utf8_1.2.1        munsell_0.5.0     crayon_1.4.1      zoo_1.8-9
```
