### 2021-08-11 9d57edb
edits for release 6.0
