+++ package/tests/testthat/test-build.R
@@ -1,8 +1,8 @@
-test_that("map constructors work", {
+test_that("map.build constructors work", {
   data(iris)
   df<-subset(iris,select=-Species)
   labels<-subset(iris,select=Species)
-  m<-map(df,labels,seed=42)
+  m<-map.build(df,labels,seed=42)
 
   expect_equal(m$convergence < 0.8,TRUE)
   expect_equal(length(m$unique.centroids)<5,TRUE)
