+++ package/tests/testthat/test-summary.R
@@ -1,11 +1,11 @@
-test_that("summary function works", {
+test_that("map.summary function works", {
   data(iris)
   df<-subset(iris,select=-Species)
   labels<-subset(iris,select=Species)
-  m<-map(df,labels,seed=42)
+  m<-map.build(df,labels,seed=42)
 
   # run summary and pull out convergence from the report tables
-  s<-summary(m,verb=FALSE)
+  s<-map.summary(m,verb=FALSE)
   conv<-s$quality.assessments[["convergence"]]
 
   expect_equal(conv < 0.8,TRUE)
