### 2021-08-11 6cefaf4
edits addressing #12 and #9
