### 2021-08-11 #3524 @lutzhamel
@HerrMo Thank you for your insightful and detailed review. Addressing the issues you raised certainly made the package a better package.  I have addressed all the issues you raised above (see the repo for detailed comments).  In the process I had to create a new version of popsom : v 6.0.  This version is now available in the repository.  I have not pushed it out to CRAN yet just in case there are additional items I need to address.  You can build the package locally on your machine by following the instructions [here](https://github.com/lutzhamel/popsom/blob/master/BUILD_RELEASE.md).  The R `check` procedure builds a package locally which you can install and run in your R environment.  Thanks again and please let me know if there is anything else I need to address.
