### 2021-08-11 #3524 @lutzhamel
@HerrMo @fabian-s  I noticed that the only bullet point that has not been not checked off in @HerrMo  review is the `automated tests` point.  You can find information on how to execute automated tests for popsom [here](https://github.com/lutzhamel/popsom/blob/master/BUILD_RELEASE.md).
