### 2021-08-12 #3524 @fabian-s
thanks @lutzhamel, we're still waiting for @rcannood to chime in.
in the meantime, could you reduce the sizes of figures 1 and 2 so they fit on one page?
