### 2021-08-12 cd31f61
minor edits to paper
