### 2021-08-12 aa7be27
minor edits to paper
