+++ joss/paper.md
@@ -192,8 +192,8 @@ The output of the script run on an Intel Linux machine is (slightly edited for r
 2     som        33.9
 3 kohonen      1975.9
 ```
-This output shows that `popsom` is about an order of magnitude
-faster than `som` and almost two orders of magnitude faster than the `kohonen` package. A
+This output shows that our `popsom` package is about an order of magnitude
+faster than the `som` package [@som2016R] and almost two orders of magnitude faster than the `kohonen` package [@kohonen2018jss]. A
 more detailed performance analysis of the `popsom` package appears in [@hamel2018vsom].
 
 # Acknowledgements
