### 2021-08-12 #3524 @lutzhamel
Hello @fabian-s. I reduced the size and will generate a new pdf.
