### 2021-08-16 lutzhamel/popsom#13 @rcannood
Split up package code to improve readability
The package consists of a very large R file which makes navigation of the functions in the package difficult. Following common R practices, I suggest splitting this file up into multiple R files, ideally one file per function.

Please consider generating the man/* files using [roxygen2](https://roxygen2.r-lib.org/articles/roxygen2.html#basic-process-1). Having the documentation close to the functions allow developers to easily interpret the functionality of each function.
