### 2021-08-16 lutzhamel/popsom#14 @rcannood
Elaborate on performance claims
The authors compare the performance of their method to the som and kohonen packages.

- Authors claim measured speedups of 60× in comparison to som (Yan, 2016). However, according to the author's own time measurements, this speedup is closer to 15×. Can you provide more insight into how
- Can authors also provide insight into the scalability of the algorithm for the execution time in function of input dataset size (number of observations and number of features).
