### 2021-08-16 lutzhamel/popsom#15 @rcannood
Provide a high-level description of SOM
Summary does not contain a high-level description of the package functionality other than that it provides an implementation of self-organising maps and that a self-organising map is an artificial neural network designed for unsupervised learning. This is insufficient information to get a high-level idea of what SOMs are and what they are commonly used for.
