### 2021-08-16 lutzhamel/popsom#16 @rcannood
Statement of need is missing
The authors statement of need is more of a results section. The only statement of need in this paragraph is: "Training a self-organizing map is time consuming." Authors should expand on how training SOMs is time consuming: which application domains? what is the dimensionality of such datasets resulting in the slow execution time?
