### 2021-08-16 lutzhamel/popsom#17 @rcannood
Improve quality of writing
Paragraphs should consist of full sentences. Partial sentences are used when referring to pieces of code or when citing articles.

Example of incomplete sentence due to code block:

> This is easily verified with a scatter plot matrix of the iris dataset using
> -code block-
> and shown in Figure 2.

Example of incomplete sentence due to citation:

> A number of R-packages exist that implement self-organizing maps including (Wehrens & Kruisselbrink, 2018) and (Yan, 2016).

If you omit everything between parentheses, the sentence that remains is incomplete:

> A number of R-packages exist that implement self-organizing maps including and.
