### 2021-08-16 lutzhamel/popsom#18 @rcannood
Reduce self-citations and refer to relevant work
The references in the manuscript fall in one of following categories:
* author's own publications (Hamel 2018, Hamel 2016, Hamel 2011, Hamel 2021, Tatoian 2018)
* publications on which the author's work is directly based (Kohonen 2001, Ultsch 1990)
* publications on which the author's work is directly compared to (Yan 2016, Wehrens 2018)
* references to applications of SOM in the field (Liakos 2018, Mathys 2019, Matić 2018, Miller 1996)

Please update the manuscript to address the following issues:
- Reduce the number of self-citations to <25%. Currently 38% of citations are self-citations-
- The citations which are meant to give a broad overview of the applications of SOM in the field are very superficial: "Self-organising maps have been applied in virtually every scientific discipline where some sort of data exploration or analysis is necessary." This is a very broad statement which is hard to verify. Please elaborate a bit on common applications of SOM. For example, FlowSOM uses SOMs to analyse Flow Cytometry data to gain insight into the expression profiles of smaller cell populations which might otherwise be missed (Van Gassen et al, Cytometry A, 2015).
