### 2021-08-16 #3524 @rcannood
Sorry for taking a while to respond. After getting back from holidays, the last few weeks have been quite hectic and I wanted to make sure I took sufficient time to properly go through the manuscript and the code before ticking off any boxes.

I appreciate the valuable contribution @lutzhamel has made in providing an efficient implementation of SOM in R. However, I feel that a few issues need to be resolved before I can approve this manuscript for publication.

I will include the first to in this comment, as it is more related to the editorial process.

## Contribution and authorship

Michael Eiger (@Meiger00) also created a significant number of commits to the repository. Should Michael also be an author on the manuscript?
Similarly, the package description lists even more authors. Should these also be included on the manuscript as well?

## Substantial scholarly effort

The description paragraph mentions that the popsom package has a number of distinguishing features, most of which are already part of a separate publication by the author (Hamel 2018, Hamel&Brown 2011, Hamel 2016 and Tatoian&Hamel 2018). I believe this publication is valuable because it could provide a nice summary of the package functionality in an open journal (whereas some of the aforementioned publications are not). In light of this, can @lutzhamel please clarify what the [substantial scientific effort](https://joss.readthedocs.io/en/latest/submitting.html#co-publication-of-science-methods-and-software) of this submission is?

## Major issues

* #lutzhamel/popsom#13
* #lutzhamel/popsom#14
* #lutzhamel/popsom#15
* #lutzhamel/popsom#16
* #lutzhamel/popsom#17
* #lutzhamel/popsom#18
