### 2021-08-16 lutz-hamel/popsom#14 @lutzhamel
I refer to a paper with a detailed performance analysis.  I don't think it is appropriate in this 'software paper' to go into such detail.
