### 2021-08-16 lutz-hamel/popsom#18 @lutzhamel
Well, to me this seems that you are just citing additional special cases...where is the cutoff?  I decided to give a high-level sketch of things that folks have applied SOMs to which I think is sufficient to show that variety of fields SOMs have been applied to.  Kohonen's latest incarnation of his book cites 1000+ papers applied to SOMs.
