### 2021-08-16 lutz-hamel/popsom#17 @lutzhamel
I don't understand what is wrong with the first problem raised in this issue; in software papers references to code are considered to be part of the prose.  For the second problem, it is standard practice to have citations to be part of the prose in technical publications.
