### 2021-08-16 lutz-hamel/popsom#14 @rcannood
Sorry @lutzhamel, but I can't agree with you here. The JOSS guidelines ask me to verify:

> If there are any performance claims of the software, have they been confirmed? (If there are no claims, please check off this item.)

The claim is that popsom is 60× faster than current SOM implementations in R, which it might be. However, based on the results you included in the manuscript, it seems like it closer to 15× than to 60×. If you'd like to convince me popsom is 60× faster, I'd be happy to take a look at these results.
