### 2021-08-16 #3524 @lutzhamel
I do not agree with any of these criticisms and hereby **retract my paper**. My understanding is that JOSS papers are supposed to be short introductions to the software packages (according to the spec in the JOSS website) these criticisms want to make this paper into an academic treatment which is not the goal of these software papers.  All of these issues have been addressed in my publications (and therefore there are many self-references) except the "quality of writing" which I find kind of insulting as well as the "split up the package" issue which I find completely strange.  This package predates roxygen and I see no benefit from retrofitting my code to this tool.
