### 2021-08-16 #3524 @lutzhamel
@fabian-s I don't agree with any of the above criticisms of my paper and therefore I am *retracting my paper*. According to the JOSS specs, JOSS papers are supposed short introductions to software packages, these criticism are trying to turn this into an academic treatment.  All these issues have been addressed in my publications (hence the number of self-citations), except "improve quality of writing" which I find kind of insulting and the "split up package" which I find bizarre.  This package predates `roxygen` and I don't see any benefit of retrofitting my package to this tool.
