### 2021-08-16 #3524 @rcannood
Please do not interpret my comments as harsh criticisms, as I only intended to improve the readability of the package and the manuscript. If you do not agree with certain comments, I kindly invite you to simply respond to my comment, as you have done for several of my comments. For instance:

> This package predates roxygen and I see no benefit from retrofitting my code to this tool.

I saw and understood that this tool predates roxygen, which is why I did not make a separate GitHub issue and simply noted "Please consider generating the man/* files using roxygen2."
