### 2021-08-16 lutz-hamel/popsom#13 @fabian-s
I fully agree with the first part (multiple thematically focussed files are better than one huge file), but if  @lutzhamel does not want to do this, it won't block acceptance. Same for roxygenizing the docs.
