### 2021-08-16 lutz-hamel/popsom#14 @fabian-s
See also https://github.com/lutzhamel/popsom/issues/9#issue-964151813.

In my view, the issue here is that the paper claims "Up to 60x faster than traditional single-threaded training algorithms", based on the experiments in Hamel 2018 in the first section and then provides actual benchmark times for R at the very end of the paper, which support the claims made in the text that "popsom package is about an order of magnitude faster than the som package (Yan, 2016) and almost two orders of magnitude faster than the kohonen package (Wehrens & Kruisselbrink, 2018)"

I would suggest to either

a) resolve the partial/apparent contradiction between these two claims by making the claim  "Up to 60x faster than traditional single-threaded training algorithms" more precise -- just add a phrase summarizing which algorithm(s) exactly this was compared to in the cited paper.
or

b) foreshadow the actual R benchmark results created for this paper in the Iitroduction *instead of* referencing these earlier comparisons in Hamel 2018 that seem less relevant to the R software that is described in the paper anyways.
