### 2021-08-16 lutz-hamel/popsom#15 @fabian-s
In my view, this is a reasonable request to improve the readability of the paper.
A 2-3 sentence summary of what SOMs do would help readers understand how and when to use this software.
