### 2021-08-16 lutz-hamel/popsom#16 @fabian-s
I agree with @rcanood that the statement of need could focus more on the gaps in available software implementation that popsom fills and less on what popsom itself does, but the two are obviously strongly linked.

The two implicit statement of needs being made here are:
- usability of existing packages suffers from computations being too slow and popsom improves that by being 1 to 2 magnitudes faster
- SOMs are mostly not very useful without good visualisations so popsom includes an improved viz method that is not available from any other package.

As far as algorithmic efficiency is concerned, I do think it would be helpful to talk about the scaling behavior of popsom in the number of data points and their dimension and compare that to alternatives. I don't think it would necessarily be very helpful to list application domains or data set sizes, there are some timings at the end of the paper for toy datasets, after all.

For the viz aspect, I think the current phrasing is fine.
