### 2021-08-16 lutz-hamel/popsom#17 @fabian-s
In my view, the first part is fine as is.

For the second part, you should use references without parentheses since these are the objects of your sentence, i.e., plz change it to "A number of R-packages exist that implement self-organizing maps including Wehrens & Kruisselbrink (2018) and Yan (2016)." This is a purely cosmetic change and should be very easy to achieve.
