### 2021-08-16 lutz-hamel/popsom#18 @fabian-s
These citations seem fine to me, a large ratio of self-citations is unavoidable in a software paper that implements the author's own work.
