### 2021-08-16 #3524 @fabian-s
@lutzhamel very sorry to hear that --
in my view, @rcannood has provided much valuable feedback and the few remaining issues that are blocking acceptance as opposed to mere suggestions for possible improvements to your code or the paper should be straightforward to adress, I think.
I've added comments to the respective issues, I think we can work this out fairly easily if you're still willing.
