### 2021-08-16 #3524 @lutzhamel
@fabian-s From your comments it seems that I have misinterpreted some of the issues raised by @rcannood, for example, "provide an explanation of how SOM's work" I took as a detailed explanation of SOMs which clearly would be beyond the scope of a short software paper.  Your interpretation as a brief sentence or two as a way to provide context to readers not familiar with the topic seems reasonable to me.  My apologies to @rcannood if I misunderstood the intention of those comments.  I do appreciate the time you took to carefully review the work!

Given this, I agree with @fabian-s  that the paper can be brought up to par without a lot of additional verbiage and work.  A recurring  theme in the reviews is the performance. I redid the performance analysis from 3 years ago and at this point can no longer reproduce the 60x speedup.  The highest speedup I can achieve now is 20x.  There are most likely two reasons for that: (1) in the past two years or so the unix world has switched from gcc to clang as its default compiler which most likely has many more aggressive optimizations based on llvm than gcc  and this diminished the performance gap between fortran and c (2) hardware has progressed, larger caches, faster clocks which might have had a disproportional impact on C code compared to Fortran code.  I will adjust the claims in the paper and the software package to bring them in line with the current findings.

The remaining issues I will address in the coming days.

Regards.
