### 2021-08-18 lutz-hamel/popsom#13 @lutzhamel
The software package has been stable for the better part of a decade.  At this point I am not inclined to rearchitect its software structure.
