### 2021-08-18 lutz-hamel/popsom#14 @lutzhamel
Agreed, I can no longer reproduce the original numbers from (Hamel, 2018). I am creating a new set of benchmarks highlighting expected speedups and possible limitations.
