### 2021-08-18 lutz-hamel/popsom#16 @lutzhamel
Rewrote the part of the statement of need that talks about speedups.
