### 2021-08-18 lutz-hamel/popsom#15 @lutzhamel
Rewrote the 'Background' section to include a high-level statement of what SOMs do.
