### 2021-08-18 lutz-hamel/popsom#17 @lutzhamel
I do not know how to control whedon's formatting of citations so I hope I achieved the same thing by rewriting the sentence slightly.
