### 2021-08-18 #3524 @lutzhamel
I will address the points raised by @rcannood [outside of the popsom repository](https://github.com/openjournals/joss-reviews/issues/3524#issuecomment-899304603):

>Contribution and authorship
>Michael Eiger (@Meiger00) also created a significant number of commits to the repository. Should Michael also be an author on the manuscript?
>Similarly, the package description lists even more authors. Should these also be included on the manuscript as well?

I am the sole author of this paper which I consider to be different and distinct of the authorship of package which includes a number of folks.  However, I am also the main contributor, architect, and maintainer of the package. I do acknowledge the contributions to the package in the paper.  I suppose, ultimately it is an editorial decision of how precisely authorship of the paper should be represented.

>Substantial scholarly effort
>The description paragraph mentions that the popsom package has a number of distinguishing features, most of which are already part of a separate publication by the author (Hamel 2018, Hamel&Brown 2011, Hamel 2016 and Tatoian&Hamel 2018). I believe this publication is valuable because it could provide a nice summary of the package functionality in an open journal (whereas some of the aforementioned publications are not). In light of this, can @lutzhamel please clarify what the substantial scientific effort of this submission is?

All the papers that @rcannood mentions in the above paragraph focus on various, distinct facets of our package. This paper pulls all these different strands of research together.  Substantial scientific effort has been expended over the better part of the last decade developing this software to the point where it is now and this paper presents the package as a whole rather than any specific aspect of it.
