### 2021-08-18 lutz-hamel/popsom#13 @rcannood
Ok with me.

If you do ever come back to separating the functions into separate files (which should not be that much work), converting the 'man/*' files to roxygen2 documentation is quite straightforward if you use `Rd2roxygen::Rd2roxygen('path/to/source/pkg')`.
