### 2021-08-18 #3524 @rcannood
In response to your earlier message: I apologise for some of the phrasing used in my initial comments; I think I could have done a better job of communicating the constructive part of my messages without coming across as critical as I did.

In response to your last message: Thanks for clarifying the matters regarding authorship and scholarly effort. Note that I asked about this specifically because whedon asked me to look into this.

Looking forward to the updated manuscript!
