### 2021-08-18 lutz-hamel/popsom#17 @rcannood
[This link](https://pandoc.org/MANUAL.html#citations-in-note-styles) might be what you need to drop the parentheses.
