### 2021-08-18 lutz-hamel/popsom#16 @fabian-s
@lutzhamel sorry if I'm overlooking something but I don't think these changes are showing up in your GH repo yet -- the last commit is 6d old.
Could you try to push all your latest commits so we can generate the updated paper?
