### 2021-08-18 07b9809
edits to address issues #14 #15 #16 #17
