+++ joss/paper.md
@@ -1,5 +1,5 @@
 ---
-title: 'popsom: A Very Efficient Implementation of Self-Organizing Maps with Starburst Visualizations for R'
+title: 'popsom: An Efficient Implementation of Self-Organizing Maps with Starburst Visualizations for R'
 tags:
   - R
   - machine learning
@@ -19,24 +19,26 @@ bibliography: paper.bib
 # Background
 
 The self-organizing map (SOM) was developed by Teuvo Kohonen for data exploration and
-data visualization in the 1980s [@kohonen2001self]. It is an artificial neural
-network designed for unsupervised learning.  What makes the self-organizing map so attractive are the intuitive mathematical
+data visualization in the 1980s [@kohonen2001self]. Specifically, the self-organizing map
+is an unsupervised artificial neural network designed for identifying and visualizing clusters in multi-dimensional
+spaces.  What makes the self-organizing map so attractive are the intuitive mathematical
 underpinnings and the straightforward visualization of computational
-results [@ultsch1990self]. Self-organizing maps have been applied in virtually every scientific discipline where some sort of data exploration or analysis is
+results [@ultsch1990self]. Self-organizing maps have been applied in virtually every
+scientific discipline where some sort of data exploration or analysis is
 necessary, e.g., [@liakos2018machine; @mathys2019single; @matic2018interpreting;
-@miller1996star].  A number of R-packages exist that implement self-organizing maps including
-[@kohonen2018jss] and [@som2016R].
+@miller1996star].  A number of R-packages exist that implement self-organizing maps, e.g.,
+[@kohonen2018jss; @som2016R].
 
 
 # Statement of need
 
 Training a self-organizing map is time consuming. Here
 we introduce an R-package called `popsom` [@popsom2021R] that implements a training algorithm
-for self-organizing maps based on vector and matrix operations inspired by
-tensor algebra [@hamel2018vsom].  We have measured speed ups of the training phase
-of a SOM  of up to 60 times using our implementation over traditional implementations of the training algorithm such as
-[@som2016R].  This speedup enables researchers to look at much larger data sets or to improve
-their throughput with a given data size.
+for self-organizing maps inspired by
+tensor algebra [@hamel2018vsom] which provides significant speedups over
+traditional implementations of the training algorithm such as the `som` package by Yan
+[@som2016R].  These speedups enable researchers to look at much larger datasets or to improve
+their throughput with a dataset of a given size.
 
 Visualization of a trained map is at the core of using self-organizing maps.  The
 `popsom` package improves on the standard u-matrix visualization
@@ -48,7 +50,8 @@ highlight cluster structures [@hamel2011improved].
 At a slightly more detailed level, our `popsom` package implements
 Kohonen's self-organizing maps with a number of distinguishing features:
 
-1. A very efficient, single threaded, stochastic training algorithm based on ideas from tensor algebra.  Up to 60x faster than traditional single-threaded training algorithms. No special accelerator hardware is required.  The speedup results from the fact that the
+1. A very efficient, single threaded, stochastic training algorithm based on ideas from tensor algebra providing significant speedups over traditional single-threaded implementations
+without the need for special accelerator hardware. The speedups result from the fact that the
 vector and matrix structures exposed by our algorithm map neatly into
 vector and matrix operations available on today's CPUs.  Our Fortran 90 implementation
 insures that these vector and matrix operations are mapped onto the hardware as efficiently as
@@ -56,7 +59,7 @@ possible [@hamel2018vsom].
 
 2. Automatic centroid visualization and detection using starbursts [@hamel2011improved]. Not
 only does `popsom` display clusters and centroids on the map using starbursts but it
-also computes a cluster model similar to a k-means model based on the starbursts.  
+also computes a cluster model based on the starbursts.  
 
[…]
