### 2021-08-18 #3524 @lutzhamel
@fabian-s @rcannood @HerrMo I have generated a new PDF and committed all my changes to the repo.  Please take a look and let me know what you think.  Thanks!
