+++ package/DESCRIPTION
@@ -1,6 +1,6 @@
 Package: popsom
 Version: 6.0
-Title: A Very Efficient Implementation of Kohonen's Self-Organizing Maps (SOMs) with Starburst Visualizations
+Title: An Efficient Implementation of Kohonen's Self-Organizing Maps (SOMs) with Starburst Visualizations
 Authors@R: c(person("Lutz", "Hamel", role = c("aut", "cre"), email = "lutzhamel@uri.edu"),
              person("Benjamin", "Ott", role = "aut"),
              person("Gregory", "Breard", role = "aut"),
@@ -8,10 +8,10 @@ Authors@R: c(person("Lutz", "Hamel", role = c("aut", "cre"), email = "lutzhamel@
              person("Michael", "Eiger", role = "aut"),
              person("Vishakh", "Gopu", role = "aut"))
 Imports: fields, graphics, ggplot2, hash, stats, grDevices
-Suggests: 
+Suggests:
     testthat (>= 3.0.0)
 Description: Kohonen's self-organizing maps with a number of distinguishing features:
-    (1) A very efficient, single threaded, stochastic training algorithm based on ideas from tensor algebra.  Up to 60x faster than traditional single-threaded training algorithms. No special accelerator hardware required (see <doi:10.1007/978-3-030-01057-7_60>).
+    (1) An efficient, single threaded, stochastic training algorithm based on ideas from tensor algebra.  Provides significant speedups over traditional single-threaded training algorithms. No special accelerator hardware required (see <doi:10.1007/978-3-030-01057-7_60>).
     (2) Automatic centroid detection and visualization using starbursts.
     (3) Two models of the data: (a) a self organizing map model, (b) a centroid based clustering model.
     (4) A number of easily accessible quality metrics for the self organizing map and the centroid based cluster model (see <doi:10.1007/978-3-319-28518-4_4>).
