### 2021-08-19 #3524 @lutzhamel
@fabian-s Problems! I was doing my final due diligence on the  paper which included carefully reviewing the performance data when I noticed that I had misunderstood a parameter setting on the `kohonen` package.  After fixing that mistake and rerunning the benchmarks I got the following results,

```
Iris Data
  package        mean time          speedup
1  popsom       215.107746                1
2     som 4074.47307966667  18.941545134626
3 kohonen 312.952335666667 1.45486316269925
```

```
Epil Data
  package        mean time          speedup
1  popsom        399.20453                1
2     som      4731.625202 11.8526340420035
3 kohonen 456.656295666667 1.14391561555343
```

```
Wines Data
  package        mean time          speedup
1  popsom       622.578688                1
2     som      5463.562089 8.77569726415049
3 kohonen 651.237032333333 1.04603168223666
```

```
Synthetic Data
  package        mean time          speedup
1  popsom 15.4396670093333                1
2     som 163.425318395667 10.5847696259819
3 kohonen 16.8341376496667 1.09031740383328
```

which implies that the `kohonen` package is pretty much on par with our `popsom` package performance wise.

What are my options?  I could withdraw the paper because one of its claims is no longer true.  I could rewrite the paper with a concentration on ease of use and interpretability of results which was one of the major angles of our research besides performance.

Please let me know how you would like to proceed.

Kind Regards,
Lutz
