### 2021-08-20 #3524 @fabian-s
@lutzhamel
hmm.... This is only the 2nd submission I'm handling for JOSS, so I don't feel comfortable deciding this myself.
I'll discuss with the JOSS team and get back to you next week.
