### 2021-08-24 #3524 @fabian-s
@lutzhamel

Considering the substantial visualization capabilities and extensive documentation of `kohonen`, we think it's unlikely that such a rewrite would demonstrate a sufficiently substantial contribution to the available open source software in this area or that it is likely to be used and cited widely.

My recommendation would be to withdraw the paper at this point.
