### 2021-08-24 #3524 @danielskatz
@fabian-s - for anything AEiC related, please ping @openjournals/joss-eics in the future, rather than a specific person
