+++ joss/paper.bib
@@ -77,7 +77,7 @@
 }
 
 @Manual{popsom2021R,
-    title = {popsom: A Very Efficient Implementation of Kohonen's Self-Organizing
+    title = {popsom: AnEfficient Implementation of Kohonen's Self-Organizing
 Maps (SOMs) with Starburst Visualizations},
     author = {Lutz Hamel and Benjamin Ott and Gregory Breard and Robert Tatoian and Michael Eiger and Vishakh Gopu},
     year = {2021},
