### 2021-12-20 6a3a791
updated the 'BUILD_RELEASE.md' doc
