### 2021-12-20 a666dcf
refined the direct package install description
